refactor: rename github datasource to github-tags
