+++ lib/datasource/github-tags/index.ts
@@ -4,7 +4,7 @@ import { logger } from '../../logger';
 
 const { get: ghGot } = api;
 
-const cacheNamespace = 'datasource-github';
+const cacheNamespace = 'datasource-github-tags';
 function getCacheKey(repo: string, type: string): string {
   return `${repo}:${type}`;
 }
