+++ lib/datasource/go/index.ts
@@ -1,10 +1,10 @@
 import { logger } from '../../logger';
 import got from '../../util/got';
-import * as github from '../github';
+import * as github from '../github-tags';
 import { DigestConfig, PkgReleaseConfig, ReleaseResult } from '../common';
 import { regEx } from '../../util/regex';
 import {
-  DATASOURCE_GITHUB,
+  DATASOURCE_GITHUB_TAGS,
   DATASOURCE_GO,
 } from '../../constants/data-binary-source';
 
@@ -17,15 +17,18 @@ async function getDatasource(name: string): Promise<DataSource | null> {
   if (name.startsWith('gopkg.in/')) {
     const [pkg] = name.replace('gopkg.in/', '').split('.');
     if (pkg.includes('/')) {
-      return { datasource: DATASOURCE_GITHUB, lookupName: pkg };
+      return { datasource: DATASOURCE_GITHUB_TAGS, lookupName: pkg };
     }
-    return { datasource: DATASOURCE_GITHUB, lookupName: `go-${pkg}/${pkg}` };
+    return {
+      datasource: DATASOURCE_GITHUB_TAGS,
+      lookupName: `go-${pkg}/${pkg}`,
+    };
   }
   if (name.startsWith('github.com/')) {
     const split = name.split('/');
     const lookupName = split[1] + '/' + split[2];
     return {
-      datasource: DATASOURCE_GITHUB,
+      datasource: DATASOURCE_GITHUB_TAGS,
       lookupName,
     };
   }
@@ -44,7 +47,7 @@ async function getDatasource(name: string): Promise<DataSource | null> {
       logger.debug({ depName: name, goSourceUrl }, 'Go lookup source url');
       if (goSourceUrl && goSourceUrl.startsWith('https://github.com/')) {
         return {
-          datasource: DATASOURCE_GITHUB,
+          datasource: DATASOURCE_GITHUB_TAGS,
           lookupName: goSourceUrl
             .replace('https://github.com/', '')
             .replace(/\/$/, ''),
@@ -85,7 +88,7 @@ export async function getPkgReleases({
 }: Partial<PkgReleaseConfig>): Promise<ReleaseResult | null> {
   logger.trace(`go.getPkgReleases(${lookupName})`);
   const source = await getDatasource(lookupName);
-  if (source && source.datasource === DATASOURCE_GITHUB) {
+  if (source && source.datasource === DATASOURCE_GITHUB_TAGS) {
     const res = await github.getPkgReleases(source);
     if (res && res.releases) {
       res.releases = res.releases.filter(
@@ -112,7 +115,7 @@ export async function getDigest(
   value?: string
 ): Promise<string | null> {
   const source = await getDatasource(lookupName);
-  if (source && source.datasource === DATASOURCE_GITHUB) {
+  if (source && source.datasource === DATASOURCE_GITHUB_TAGS) {
     // ignore v0.0.0- pseudo versions that are used Go Modules - look up default branch instead
     const tag = value && !value.startsWith('v0.0.0-2') ? value : undefined;
     const digest = await github.getDigest(source, tag);
