+++ lib/manager/nvm/extract.ts
@@ -1,12 +1,12 @@
 import { isValid } from '../../versioning/node';
 import { PackageFile, PackageDependency } from '../common';
-import { DATASOURCE_GITHUB } from '../../constants/data-binary-source';
+import { DATASOURCE_GITHUB_TAGS } from '../../constants/data-binary-source';
 
 export function extractPackageFile(content: string): PackageFile {
   const dep: PackageDependency = {
     depName: 'node',
     currentValue: content.trim(),
-    datasource: DATASOURCE_GITHUB,
+    datasource: DATASOURCE_GITHUB_TAGS,
     lookupName: 'nodejs/node',
   };
   if (!isValid(dep.currentValue)) {
